net: dsa: ksz9477: use a kthread work item per deferred skb
There might be a race in tag_ksz.c between these two lines:

	skb_queue_tail(&ptp_shared->xmit_queue, skb_get(skb));
	kthread_queue_work(ptp_shared->xmit_worker, &ptp_shared->xmit_work);

and the skb dequeue logic in ksz9477_port_deferred_xmit(). For example,
the xmit_work might be already queued, however the work item has just
finished walking through the skb queue. Because we don't check the
return code from kthread_queue_work, we don't do anything if the work
item is already queued.

However, nobody will take that skb and send it, at least until the next
timestampable skb is sent.

With the ksz9477 driver, two-step TX timestamping is a rare process, and
in certain configs it may happen even as rarely as once per second.

So if the race condition described above happens, we might experience
huge delays.

To close that race, let's not keep a single work item per port, and a
skb timestamping queue, but rather dynamically allocate a work item per
packet.

It is also unnecessary to have more than one kthread that does the work.
So delete the per-port kthread allocation and replace them with a single
kthread which is global to the switch.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>