net: switchdev: export switchdev_lower_dev_find
This little function that retrieves the first lower interface of @dev
that passes the @check_cb and @foreign_dev_check_cb criteria is useful
outside of the switchdev core, too. For example, drivers may use it to
retrieve a pointer to one of their own netdevices beneath a LAG.

Export it for driver use, to reduce code duplication.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>