net: dsa: move dsa_foreign_dev_check above dsa_slave_switchdev_event_work
In preparation of LAG FDB support, we'll need to call
switchdev_lower_dev_find in order to get a handle of a DSA user port
from a LAG interface. That function takes dsa_foreign_dev_check() as one
of its arguments. So to avoid forward declarations,
dsa_foreign_dev_check() needs to be moved above it.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>