net: dsa: sja1105: update virtual links when a port enters or exits a bridge
After change "net: dsa: tag_8021q: replace the SVL bridging with
VLAN-unaware IVL bridging", tag_8021q enforces two different pvids on a
port, depending on whether it is standalone or in a VLAN-unaware bridge.

Up until now, there was a single pvid, represented by
dsa_tag_8021q_rx_vid(), and that was used as the VLAN for VLAN-unaware
virtual link rules, regardless of whether the port was bridged or
standalone.

To keep VLAN-unaware virtual links working, we need to follow whether
the port is in a bridge or not, and update the VLAN ID from those rules.

In fact we can't fully do that. Depending on whether the switch is
VLAN-aware or not, we can accept Virtual Link rules with just the MAC
DA, or with a MAC DA and a VID. So we already deny changes to the VLAN
awareness of the switch. But the VLAN awareness may also change as a
result of joining or leaving a bridge.

Therefore, just allow the following: a port may leave a VLAN-unaware
bridge while it has VLAN-unaware VL (tc-flower) rules, and the driver
will update those with the new tag_8021q pvid for standalone mode, but
the driver won't accept joining a bridge at all while VL rules were
installed in standalone mode. This is sort of a compromise made because
leaving a bridge is an operation that cannot be vetoed. But this sort of
setup change is not fully supported, either: as mentioned, VLAN
filtering changes can also be triggered by leaving a bridge, therefore,
the existing veto we have in place for turning VLAN filtering off with
VLAN-aware VL rules active still isn't fully effective. But nonetheless,
this is the best we can hope for.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>