irqchip/slir: add driver for NXP SJA1110 interrupt controller for embedded PHYs
The SJA1110 automotive Ethernet switch has an INT_N line towards the
host SoC and an internal interrupt controller which multiplexes a bunch
of internal signals onto that single IRQ line. These signals mostly come
from the link change interrupt of the embedded Ethernet PHYs (100BASE-T1
and 100BASE-TX).

Creating an IRQ domain for this interrupt controller allows us to tell
phylib to use interrupts for link status for the internal PHYs, which
will eliminate periodic polling of their MDIO status registers.

I've googled the register names and I found a bunch of blocks in other
hardware which seem to have the same concept behind them, so I thought a
generic driver might be useful. In the current presentation, the driver
expects to find a regmap in its parent device using devres resource 0.
However, it can easily be modified to be MMIO-driven as well (a function
such as ocelot_regmap_from_resource() can be renamed to something more
generic like platform_device_regmap_from_resource() and that can be
called directly). I didn't do that because I don't need to support MMIO
(yet).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>