net: dsa: sja1105: rename sja1105_spi.c to sja1105_soc.c
This is a preparation patch for splitting out a parent SoC driver (which
handles regmap creation and register access) out of the big DSA driver.
Currently sja1105_spi.c contains most of the logic which will be needed
by the SoC driver, hence the bulk rename of the file (to minimize the
delta in future changes).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>