net: dsa: sja1105: move static config stuff out of sja1105_soc.c
The static config processing and uploading is about the only stuff from
sja1105_soc.c which has to do with the switching IP rather than with the
SoC by and large. After the renaming from sja1105_spi.c to sja1105_soc.c,
this code is now misplaced, so move it to sja1105_main.c.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>