net: dsa: sja1105: split regs from priv->info
Currently, struct sja1105_regs is something which the SoC driver
(sja1105_soc.c) should own, and struct sja1105_info is something which
the DSA driver (sja1105_main.c) should own. But they are intertwined, we
need to split them in order to be able to probe in a split driver
configuration.

For the case where the DSA driver is probing on an old-style device
tree (DSA is in control of the spi_device, and the SoC driver is just
library code), it still needs to be able to probe on the sja1105_regs
structure.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>