net: dsa: sja1105: add platform driver for switching IP
Add a glue driver for the DSA logic, which allows the switching IP to
probe as a child of the SoC driver with new-style device trees, rather
than probing on the SPI device on old-style ones.

We use the same compatible strings as in the case of the DSA driver over
the SPI device, but the of_device_id has no data on which we can run
of_device_get_match_data(). This is because the "regs" are provided by
soc->regs, where "soc" is the dev_get_drvdata() of the parent. So there
is a built-in assumption that this platform device can only probe as a
child of the sja1105_soc driver.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>