net: dsa: replace cross-chip notifier with a list
Based on the new rule of thumb from Jakub Kicinski, that netdev
notifiers should be avoided if there is a direct way of achieving the
same result, the cross-chip notifier chain should go.

The only purpose of cross-chip notifiers, as documented, is to execute
code for each switch in a tree.

DSA has pointers to all those switches, but the tree only keeps a list
of ports, and puts ports of all switches in the same bucket, so the
cross-chip notifier mechanism was the preferred way of doing that.

If we add another list to the DSA switch tree, dst->switches, we are
able to iterate through that just fine and achieve the same effect.

The API presented by the cross-chip notifiers stays the same, even
though the implementation no longer uses notifiers.

Side note, notifiers weren't all that great for robust error teardown.
I used to have a patch set which made robust error handling possible,
but it required changes in the core notifier API, and that didn't get
too much traction or interest. In the interest of direct code
conversion, this patch attempts no graceful error handling.

Link: https://lore.kernel.org/netdev/20230406073556.7d63df19@kernel.org/
Link: https://lore.kernel.org/all/20220818154911.2973417-1-vladimir.oltean@nxp.com/
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>