net: dsa: make address deletion return void
DSA doesn't have something useful to do with the information that an
address (FDB, MDB) deletion failed - it has nowhere to propagate that
information.

Currently it tries to cling onto that address by setting its refcount
back to 1 (in case it was refcounted), and so, pretending that the
hardware driver failed cleanly and that the entry is still there,
somewhere.

Looking at some of the driver implementations, the fact that the
address is still there on failure is not immediately obvious, and
therefore, nothing is obvious.

Make DSA simply ignore errors and let drivers deal with them the best
they can.

This also modifies the prototypes of the following trace points:
trace_dsa_fdb_del_hw
trace_dsa_mdb_del_hw
trace_dsa_lag_fdb_del_hw

to stop passing an "int err" argument - the return code of
ds->ops->*_del(). And since this new prototype (without an "int err")
looks identical to the "_not_found" class of trace events, create a new
class "*del" which covers both "*_del_hw" as well as "*_del_not_found.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>