net: mscc: ocelot: make ocelot_vcap_filter_{del,replace}() return void
These return an error because the API permits calling them on struct
ocelot_vcap_filter pointers which were not added to any VCAP block, or
i.o.w. they were just kalloc()ed but no ocelot_vcap_filter_add().
So it might be possible to not find them when removing or deleting.

Changing that fundamental fact of the API would involve creating a much
larger API surface than just ocelot_vcap_filter_add(), with a truck load
of arguments, since VCAP filters can have all sorts of keys, actions and
chains. The current API gives the flexibility that the filter is
configured by the caller between the kcalloc() and the add() call.

Nonetheless, the current ocelot and felix drivers are paranoid enough to
catch buggy API usage and propagate that as errors to higher levels.
That's quite unnecessary, though, and the current usage of the API is
not buggy. Just add a WARN_ON(), exit early to not crash the kernel, and
make them return void.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>