net: dsa: make VLAN deletion return void
Currently, the things which drivers are trying to say through
ds->ops->port_vlan_del() and ds->ops->tag_8021q_vlan_del() error codes
are either that the operation isn't supported, or that hardware access
failed.

The former is moot since DSA will never delete a VLAN that was never
added (and the restriction can be added there), and the latter is not
something that DSA can do anything about - the hardware is pretty toast
at this stage.

Simplify the API and make these methods return void.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>