net: dsa: fail operation early if ageing time is out of range for any switch
DSA's algorithm of dealing with ageing time changes is to emit a
cross-chip notifier and to stop mid-way if one switch does not support
that particular value. There is no rollback attempt because the old
value is not even saved.

Now that we have dst->switches, iterate through that first, before
emitting the cross-chip "notifier".

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>