phy: introduce phy_get_status() and use it to report CDR lock
Clock and Data Recovery, or CDR, is a fundamental building block in high
speed serial interface receivers, which is a protocol-agnostic component
that uses samplers to detect bit transitions in the incoming RX stream,
and deduce timing information from them. "Lock" means that the CDR
circuit is able to see enough transitions at the right frequency to
effectively be able to understand the bit stream.

Since the CDR is protocol-agnostic, we cannot state that "CDR lock"
means we have a link (this is protocol-specific), but if we do not have
CDR lock, for sure we do not have a link.

Some modules, like the MTIP AN/LT block used as a copper backplane PHY
driver, need this extra information from the SerDes PHY as another
source of "link down" information.

Namely, the 25GBase-R PCS does not have a MDIO_CTRL1_LPOWER bit
implemented in its MDIO_MMD_PCS:MDIO_CTRL1 register. That bit is
typically set from phy_suspend() or phylink_pcs_disable() implementations,
and that is supposed to cause a link drop event on the link partner.
But here it does not happen.

By implementing the networking phylink_pcs_disable() as phy_power_off(),
we are able to actually power down the lane in a way that is visible to
the remote end. Where it is visible is the CDR lock, so we introduce
PHY_STATUS_TYPE_CDR_LOCK as an extra link indication, we are able to
detect that condition and signal it to upper layers of the network
stack.

A more high-level and generic phy_get_status() operation was chosen
instead of the more specific phy_get_cdr_lock() alternative, because I
saw this as being more in the spirit of the generic PHY API.
Also, phy_get_status() is more extensible and reusable for other
purposes as well.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Reviewed-by: Florian Fainelli <florian.fainelli@broadcom.com>