net: dsa: sja1105: split struct sja1105_soc from sja1105_private
For the SoC driver to be able to work in a split configuration, it needs
to not access data structures belonging to the DSA driver. Here we need
to remove the structures necessary for register access from struct
sja1105_private, and move them to a smaller struct sja1105_soc.

Because we need to work with old-style device trees as well (where the
DSA driver handles the spi_device), struct sja1105_private will also
hold a pointer to a struct sja1105_soc.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>