net: dsa: rename dsa_user_phy_connect() to dsa_user_mii_bus_phy_connect()
We plan to make DSA connect to its PHY at ndo_open() time.
dsa_user_phy_connect() sounds like a good name for that activity, which
currently takes place in dsa_user_phy_setup().

But the name is already taken by the method which performs the non-OF
based connection to an internal PHY. We could use a more descriptive
name for this function anyway, so rename it to dsa_user_mii_bus_phy_connect(),
which both aids the understanding of the code, and frees up the name.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>