lib: packing: add pack_fields() and unpack_fields()
It is not uncommon for consumer drivers to need to pack and unpack a
large number of fields to/from a buffer.

For example in the sja1105 driver, we have common packing() functions
for both the PACK and UNPACK directions, to not duplicate the bit field
definitions.

But this comes with the drawback that we cannot have const-correct
function prototypes. It would be good to have dedicated functions for
the pack() and the unpack() operation.

That pack()/unpack() split causes the very direct problem of having to
duplicate all packing() calls, for each field. The idea to avoid that,
implemented here, is to decouple the field definitions from the function
call, and just provide them as a table. The library iterates through the
fields and translates them into unpacked variants, held as members of a
structure.

There are even more things we can do better with this new API.

First, the packing API works with unpacked numbers stored as u64. But
since we need to know the offsetof() the field within the structure, we
can also ask for the sizeof_field(). Thus, we can support smaller than
u64 unpacked field sizes. This can greatly reduce the memory footprint
of the consumer driver.

Second, the error handling with packing(), but also with pack() and
unpack(), is not exactly great. As documented, drivers should catch
return codes from these library functions, only to print something, and
then discard them, because they are likely programming bugs. But it
would be better if programming bugs would not incur runtime checks that
are being made once for every single function call. The idea here is to
move the sanity checks to compilation time. This is now possible, since
the packed fields are moved to structures, and the BUILD_BUG_ON()
infrastructure can cover that with relative easy. Note: the C
preprocessor can't generate variable-length code, as would be required
for array-style definitions of struct packed_field arrays. So I went on
a little code generation spree, and I wrote explicit macros in
include/linux/packing.h for sanity-checking arrays of 1 packed field, 2
packed fields, 3 packed fields, ..., all the way to 50 packed fields.
In practice, the sja1105 driver will actually need the variant with 40
fields. This isn't as bad as it seems: feeding a 39 entry sized array
into the CHECK_PACKED_FIELDS_40() macro will actually generate a
compilation error, so mistakes are very likely to be caught by the
developer and thus are not a problem. It's just that the length of the
array is baked into the macro definition, as I couldn't write a single
generic CHECK_PACKED_FIELDS() macro.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>