net: dsa: sja1105: fix up HOSTCMD only for PACK
sja1105pqrs_common_l2_lookup_cmd_packing() has logic to derive, rather
than store, the HOSTCMD field from the values of RDWRSET, SEARCH and
VALIDENT.

The HOSTCMD is only packed into the buffer. The driver never needs to
know what was the HOSTCMD of an entry read from the switch. Nonetheless,
the code runs in both cases. We don't want to duplicate it when we
rework to separate pack() and unpack() functions. For that, be explicit
that we only expect it to appear with PACK.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>