lib: packing: fix pack_fields_m() and unpack_fields_m() macros
In u64_to_ustruct_field() and ustruct_field_to_u64(), ustruct is a
void *, and arithmetic works on that just as on an u8 *.

As opposed to those, ustruct in the macro variant has the typeof() of
the container scope, and arithmetic works way different. If ustruct is a
structure as it should be, ustruct + field->offset is actually
interpreted as the array element &ustruct[field->offset], so it shoots
way past the goal posts. Fix that.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>