net: dsa: sja1105: propagate blk_idx to sja1105_dynamic_config_poll_valid()
This function unpacks entries, and sometimes a dynamic entry is unpacked
exactly like a static one.

Currently, the linkage is that the struct sja1105_dynamic_table_ops ::
entry_packing() redirects to the struct sja1105_table_ops :: packing(),
but we'd like the latter to go away. We need a new linkage based on
blk_idx.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>