net: dsa: sja1105: stop calling static config methods directly from dynamic config
The static config packing() interface is about to change, and we need to
give it a bit more breathing room in order to not break things.

Currently we call directly methods such as
sja1105pqrs_general_params_entry_packing() when the entry_packing() of
the dynamic entry is identical to the packing() method of the static
entry.

Instead of direct function calls, we go through the table_ops pointer
corresponding to the block index blk_idx, which is the same between the
static and the dynamic config. This is accessed via priv, so
entry_packing() needs to gain this extra argument.

We know when to call the static config variant of the packing() function
because we're making entry_packing() optional. When absent, the standard
variant is needed.

Also, since we no longer directly redirect entry_packing() to packing(),
and thus we don't have a prototype compatibility to keep, we can change
the return type of entry_packing() to void.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>