net: dsa: sja1105: split cmd_packing() and entry_packing() in the dynamic config
Propagate const correctness to the dynamic config implementation, by not
using the bidirectional packing(), but separate pack() and unpack()
instead.

This does locally result in more code and is a bit messy (as the
hardware itself in this area), but is compensated globally by the use of
packed field arrays elsewhere.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>