net: dsa: transfer rtnl_trylock() from dsa_tree_change_tag_proto() to tagging_store()
netdev sysfs's relationship with rtnl_lock() is complicated due to AB-BA
lock inversion deadlocks [1], so rtnl_trylock() -> restart_syscall() is
the coding pattern to this day in sysfs methods.

We need rtnl_lock() in tagging_store() because of the dev->dsa_ptr
dereference which needs to be replaced with dsa_conduit_to_cpu_port_rtnl()
in a subsequent change.

[1]: https://lore.kernel.org/netdev/20210928125500.167943-1-atenart@kernel.org/T/
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>